requirements.txt
# FastAPI and ASGI server
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.1.0.post1
starlette==0.27.0
typing-extensions==4.8.0
anyio==3.7.1
click==8.1.7
h11==0.14.0
httptools==0.6.1
websockets==12.0
watchfiles==0.21.0

# MongoDB
motor==3.3.2
pymongo==4.6.1
dnspython==2.4.2

# Environment variables
python-dotenv==1.0.0
python-decouple==3.8

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cryptography==41.0.5
pycryptodome==3.19.0
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6

# Twilio for WhatsApp integration
twilio==8.10.0
phonenumbers==8.13.25

# Date and time handling
python-dateutil==2.8.2
pytz==2023.3
tzlocal==5.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
pytest-cov==4.1.0
coverage==7.3.2
pytest-mock==3.12.0
pytest-env==1.1.3

# Development tools
black==23.10.1
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pylint==3.0.2
autopep8==2.0.4
pre-commit==3.5.0

# Logging
loguru==0.7.2
structlog==23.2.0
python-json-logger==2.0.7

# Async support
aiohttp==3.8.6
aiodns==3.1.1
aiosignal==1.3.1
async-timeout==4.0.3
frozenlist==1.4.1
multidict==6.0.4
yarl==1.9.4

# Utilities
python-slugify==8.0.1
Pillow==10.1.0
requests==2.31.0
urllib3==2.1.0
six==1.16.0
text-unidecode==1.3
python-magic==0.4.27
python-magic-bin==0.4.14

# API Documentation
swagger-ui-bundle==0.0.9
apispec==6.3.0
apispec-webframeworks==0.5.2

# Caching
redis==5.0.1
aioredis==2.0.1

# Background Tasks
celery==5.3.4
flower==2.0.1

# Monitoring
prometheus-client==0.18.0
sentry-sdk==1.32.0

# File Handling
aiofiles==23.2.1

# Validation
validators==0.22.0
marshmallow==3.20.1

# Database Migrations
alembic==1.12.1
mongoengine==0.27.0

# API Rate Limiting
slowapi==0.1.8
limits==3.7.0

# WebSocket Support
websockets==12.0

# Added from the code block
pyngrok==7.0.6 